Organized client and technician fully
Pretty much everything should be in place now. I changed the account.html to index.html, and it works, so let me know what was going wrong @iunary . Made a bunch of small changes just so everything is in place with the blueprints! Now just time for the profile pic upload page!